pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "task-vectors-playground"
version = "0.1.0"
requires-python = "== 3.11.8"
description = "Playing with those task vectors"
authors = [{ name = "Donato Crisostomi", email = "donatocrisostomi@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "aiohttp==3.9.3",
    "aiohttp-retry==2.8.3",
    "aiosignal==1.3.1",
    "altair==5.3.0",
    "amqp==5.2.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.3.0",
    "anypy==0.0.6",
    "appdirs==1.4.4",
    "arrow==1.3.0",
    "asyncssh==2.14.2",
    "atpublic==4.1.0",
    "attrs==23.2.0",
    "Babel==2.14.0",
    "bandit==1.7.8",
    "beautifulsoup4==4.12.3",
    "billiard==4.2.0",
    "black==24.3.0",
    "blessed==1.20.0",
    "blinker==1.7.0",
    "boto3==1.34.79",
    "botocore==1.34.79",
    "cachetools==5.3.3",
    "celery==5.3.6",
    "cffi==1.16.0",
    "cfgv==3.4.0",
    "charset-normalizer==3.3.2",
    "clarabel==0.7.1",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "configobj==5.0.8",
    "contourpy==1.2.1",
    "coverage==7.4.4",
    "croniter==1.3.15",
    "cryptography==42.0.5",
    "cvxpy==1.4.2",
    "cycler==0.12.1",
    "datasets==3.2.0",
    "dateutils==0.6.12",
    "deepdiff==6.7.1",
    "dictdiffer==0.9.0",
    "dill==0.3.8",
    "diskcache==5.6.3",
    "distlib==0.3.8",
    "distro==1.9.0",
    "docker-pycreds==0.4.0",
    "dpath==2.1.6",
    "dulwich==0.21.7",
    "dvc==3.43.1",
    "dvc-data==3.9.0",
    "dvc-gdrive==3.0.1",
    "dvc-http==2.32.0",
    "dvc-objects==3.0.6",
    "dvc-render==1.0.1",
    "dvc-studio-client==0.20.0",
    "dvc-task==0.4.0",
    "ecos==2.0.13",
    "editor==1.6.6",
    "entrypoints==0.4",
    "fastapi==0.88.0",
    "filelock==3.13.1",
    "flake8==7.0.0",
    "flatten-dict==0.4.2",
    "flufl.lock==7.1.1",
    "fonttools==4.51.0",
    "frozenlist==1.4.1",
    "fsspec==2023.12.2",
    "ftfy==6.2.0",
    "funcy==2.0",
    "ghp-import==2.1.0",
    "gitdb==4.0.11",
    "GitPython==3.1.43",
    "grandalf==0.8",
    "gto==1.6.2",
    "h11==0.14.0",
    "h5py>=3.13.0",
    "httplib2==0.22.0",
    "huggingface-hub==0.27.1",
    "hydra-core==1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "identify==2.5.35",
    "importlib_metadata==8.6.1",
    "importlib_resources==6.4.0",
    "iniconfig==2.0.0",
    "inquirer==3.2.4",
    "ipykernel>=6.29.5",
    "ipython==8.22.0",
    "isort==5.13.2",
    "iterative-telemetry==0.0.8",
    "itsdangerous==2.1.2",
    "Jinja2==3.1.2",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "kiwisolver==1.4.5",
    "kombu==5.3.6",
    "lightning==2.0.0",
    "lightning-cloud==0.5.65",
    "lightning-utilities==0.11.2",
    "Markdown==3.6",
    "markdown-it-py==3.0.0",
    "matplotlib==3.8.4",
    "mccabe==0.7.0",
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.17",
    "mpmath==1.3.0",
    "multidict==6.0.5",
    "multiprocess==0.70.16",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "nn-template-core==0.3.2",
    "nodeenv==1.8.0",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "open-clip-torch==2.0.2",
    "ordered-set==4.1.0",
    "orjson==3.10.0",
    "osqp==0.6.5",
    "packaging==24.2",
    "pandas==2.2.1",
    "pathspec==0.12.1",
    "patsy==0.5.6",
    "pbr==6.0.0",
    "pillow==10.2.0",
    "platformdirs==3.11.0",
    "plotly==5.20.0",
    "pluggy==1.4.0",
    "pre-commit==3.7.0",
    "prompt-toolkit==3.0.43",
    "protobuf==4.25.3",
    "psutil==6.1.1",
    "pytorch-lightning==2.0.0",
    "pytorch-lightning-bolts==0.3.2.post1",
    "pytz==2024.1",
    "PyYAML==6.0.1",
    "pyyaml_env_tag==0.1",
    "qdldl==0.1.7.post1",
    "requests==2.32.3",
    "rich==13.7.1",
    "scipy==1.11.4",
    "seaborn==0.13.2",
    "setproctitle==1.3.3",
    "torch==2.0.1",
    "torchmetrics==1.0.3",
    "torchvision==0.15.2",
    "tqdm==4.67.1",
    "transformers==4.39.3",
    "typer==0.12.1",
    "urllib3==2.3.0",
    "uvicorn==0.29.0",
    "wandb==0.16.6",
    "websockets==11.0.3",
]

[tool.setuptools]
packages = ["tvp"]

[tool.uv.sources]
task-vectors-playground = { workspace = true }

[project.scripts]
evaluate = "tvp.scripts.evaluate_pipeline:main"
train = "tvp.scripts.train_router:main"

[dependency-groups]
dev = [
    "task-vectors-playground",
]
